Move Button to shared components
