+++ ui/apps/dev-server-ui/src/components/Event/SendEventModal.tsx
@@ -7,11 +7,11 @@ import {
   type KeyboardEvent,
   type SyntheticEvent,
 } from 'react';
+import { Button } from '@inngest/components/Button';
 import Editor, { useMonaco } from '@monaco-editor/react';
 import { toast } from 'sonner';
 import { ulid } from 'ulid';
 
-import Button from '@/components/Button/Button';
 import Modal from '@/components/Modal';
 import useModifierKey from '@/hooks/useModifierKey';
 import { usePortal } from '../../hooks/usePortal';
