+++ ui/apps/dev-server-ui/src/components/Modal.stories.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
+import { Button } from '@inngest/components/Button';
 import type { Meta, StoryObj } from '@storybook/react';
 
-import Button from './Button/Button';
 import Modal from './Modal';
 
 const meta = {
