+++ ui/packages/components/src/Button/index.ts
@@ -1,1 +1,2 @@
 export { Button } from './Button';
+export { getButtonColors, getButtonSizeStyles, getDisabledStyles } from './buttonStyles';
